Add wrapped tooltip to editor
